Add ability for platform admins to approve/deny a store
Stores can now be approved or denied by a platform admin through the admin's platform dashboard.

Implementing this also affected the navbar (adding a new dashboard) and added a new column (approval) to the store table.
Default approval is false. A scope was also added to the store table to return only approved stores.